# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "API MTIS - OpenAPI 3.0",
    "description" : "Practica grupal  de MTIS de la Universidad de Alicante 2022-2023.\n\n_Esta documentacion esta escrita en el Swagger Editor y con OpenAPI 3.0_\n\n`Autor: Nikita Polyanskiy, Serhii Vidernikov, Robert Kazaryan, Víctor Hernández Justamante y Ángel Tamame Pérez`",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/VHJ1_1/MTIS/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/idSeguimiento" : {
      "post" : {
        "tags" : [ "IdSeguimiento" ],
        "summary" : "Generar identificador de seguimiento",
        "operationId" : "generarIdentificador",
        "parameters" : [ {
          "name" : "RestKey",
          "in" : "header",
          "description" : "Api Key",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Petición para generar un identificador de seguimiento",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/idSeguimiento_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/idSeguimiento/{id}" : {
      "post" : {
        "tags" : [ "IdSeguimiento" ],
        "summary" : "Validar identificador de seguimiento",
        "operationId" : "validarIdentificador",
        "parameters" : [ {
          "name" : "RestKey",
          "in" : "header",
          "description" : "Api Key",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Identificadora validar",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201_1"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Response" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "description" : "Estado de la peticion",
            "example" : "Success, Bad request, etc."
          },
          "message" : {
            "type" : "string",
            "description" : "Mensaje mas concreto",
            "example" : "Se ha creado con exito, El campo no es valido, etc."
          }
        },
        "example" : {
          "message" : "Se ha creado con exito, El campo no es valido, etc.",
          "status" : "Success, Bad request, etc."
        }
      },
...
      "idSeguimiento_body" : {
        "required" : [ "pedidoId" ],
        "type" : "object",
        "properties" : {
          "pedidoId" : {
            "type" : "integer",
            "description" : "ID del pedido para el que se generará el identificador de seguimiento"
          }
        }
      },
      "inline_response_201" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "example" : "LN11543234"
          }
        },
        "example" : {
          "id" : "LN11543234"
        }
      },
      "inline_response_201_1" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "boolean",
            "example" : true
          },
          "message" : {
            "type" : "string",
            "example" : "identificador correcto"
          }
        },
        "example" : {
          "id" : true,
          "message" : "identificador correcto"
        }
      }
    },
...
  }
}